refactor: rename to client-legacy to indicate intnet of maintenance
